Add path finding function into SearchGraph
