+++ pyrobosim/pyrobosim/navigation/world_graph.py
@@ -3,12 +3,10 @@
 import time
 import itertools
 
-from .a_star import AStarGraph
 from .planner_base import PathPlannerBase
 from ..core.locations import Location
-from ..utils.motion import Path
+from ..utils.motion import Path, reduce_waypoints_polygon
 from ..utils.pose import Pose
-from ..utils.motion import reduce_waypoints_polygon
 from ..utils.search_graph import SearchGraph, Node
 
 
@@ -33,7 +31,6 @@ class WorldGraphPlannerPolygon:
         self.world = world
         self.compress_path = compress_path
 
-        self.path_finder = AStarGraph()
         self.planning_time = 0.0
         self.latest_path = Path()
 
@@ -103,7 +100,7 @@ class WorldGraphPlannerPolygon:
 
         # Find a path from start to goal nodes
         t_start = time.time()
-        waypoints = self.path_finder.plan(start, goal)
+        waypoints = self.graph.find_path(start, goal)
         # Return empty path if no path was found.
         if not waypoints:
             return self.latest_path
